refactor(notifications): move some functions from utils to notifications.py
